requirements.txt
langchain
langchain-community
langchain-chroma
langchain-core
langgraph
pypdf
chromadb
yfinance
pandas
numpy
scikit-learn
langchainhub
langchain-groq
langchain-huggingface
sentence-transformers
python-dotenv